+++ NEWS.md
@@ -1,5 +1,7 @@
 # binpackr (development version)
 
+* Fixed wrong item distribution in some cases when item sizes were greater than the bin capacity.
+
 # binpackr 0.1.0
 
 * Initial CRAN submission.